requirements.txt
pygame==2.0.1
numpy==1.20.2
Pillow==8.3.1
pygameZoom==0.0.4
